Adds charm implementation
Adds a simple charm implementation that requires an ingress relation.
When used in relation with the nginx-ingress-integrator charm, this
should result in the creation of an ExternalName Kubernetes Service
for the configured `external-name` name and create an ingress route
for it.

The charm will require the external-name config option to be set.